% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl4th-package.R
\docType{package}
\name{nfl4th-package}
\alias{nfl4th}
\alias{nfl4th-package}
\title{nfl4th: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://guga31bb.github.io/nfl4th}
}

}
\author{
\strong{Maintainer}: Ben Baldwin \email{bbaldwin206@gmail.com}

Other contributors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com} [contributor]
}

}
\keyword{internal}
