% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{add_4th_probs}
\alias{add_4th_probs}
\title{Get 4th down decision probs}
\usage{
add_4th_probs(df)
}
\arguments{
\item{df}{A data frame of decisions to be computed for.}
}
\value{
Original data frame Data frame plus the following columns added:
\describe{
first_down_prob, wp_fail, wp_succeed, go_wp, fg_make_prob, miss_fg_wp, make_fg_wp, fg_wp, punt_wp
\item{first_down_prob}{Probability of earning a first down if going for it on 4th down.}
\item{wp_fail}{Win probability in the event of a failed 4th down attempt.}
\item{wp_succeed}{Win probability in the event of a successful 4th down attempt.}
\item{go_wp}{Average win probability when going for it on 4th down.}
\item{fg_make_prob}{Probability of making field goal.}
\item{miss_fg_wp}{Win probability in the event of a missed field goal.}
\item{make_fg_wp}{Win probability in the event of a made field goal.}
\item{fg_wp}{Average win probability when attempting field goal.}
\item{punt_wp}{Average win probability when punting.}
}
}
\description{
Get various probabilities associated with each option on 4th downs (go
for it, kick field goal, punt).
}
