% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{make_2pt_table_data}
\alias{make_2pt_table_data}
\title{Get 2pt decision probs}
\usage{
make_2pt_table_data(probs)
}
\arguments{
\item{probs}{A data frame that has had \code{add_2pt_probs()} already run on it.}
}
\value{
A table showing the probabilities associated with each decision.
}
\description{
Get a table with the probabilities associated with a 2-pt decision.
}
