% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{make_table_data}
\alias{make_table_data}
\title{Get 4th down decision probs}
\usage{
make_table_data(probs)
}
\arguments{
\item{probs}{A data frame that has had \code{add_4th_probs()} already run on it.}
}
\value{
A table showing the probabilities associated with each decision.
}
\description{
Get a table with the probabilities on 4th down.
}
