% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl4th-package.R
\docType{package}
\name{nfl4th-package}
\alias{nfl4th}
\alias{nfl4th-package}
\title{nfl4th: Functions to Calculate Optimal Fourth Down Decisions in the National Football League}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to estimate outcomes of fourth down
    plays in the National Football League.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nfl4th.com/}
  \item \url{https://github.com/guga31bb/nfl4th/}
  \item Report bugs at \url{https://github.com/guga31bb/nfl4th/issues}
}

}
\author{
\strong{Maintainer}: Ben Baldwin \email{bbaldwin206@gmail.com}

Other contributors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com} [contributor]
}

}
\keyword{internal}
