% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl4th-package.R
\docType{package}
\name{nfl4th-package}
\alias{nfl4th}
\alias{nfl4th-package}
\title{nfl4th: Functions to Improve Fourth Down Decisions in the NFL}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to estimate outcomes of fourth down
    plays in the National Football League.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://guga31bb.github.io/nfl4th}
}

}
\author{
\strong{Maintainer}: Ben Baldwin \email{bbaldwin206@gmail.com}

Other contributors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com} [contributor]
}

}
\keyword{internal}
